+++ src/index.ts
@@ -1,5 +1,5 @@
 export { isInstanceOf } from './components';
 export { action, component, ignoreState, sequence, withId } from './decorators';
-export { ActionOptions, AppOptions, GlobalOptions, LogLevel } from './options';
+export { AppOptions, GlobalOptions, LogLevel } from './options';
 export { getMethods } from './utils';
 export { ReduxApp } from './reduxApp';